requirements.txt
streamlit==1.55.0
pandas==2.3.3
yfinance==1.2.0
plotly==6.6.0
google-genai